Fix license requests for PlayReady on Edge (parse message contents for headers and body)
