+++ src/ms-prefixed.js
@@ -1,38 +1,8 @@
 import videojs from 'video.js';
 import window from 'global/window';
+import { requestPlayreadyLicense } from './playready';
 
 const PLAYREADY_KEY_SYSTEM = 'com.microsoft.playready';
-
-export const getMessageContents = (message) => {
-  const xml = (new DOMParser()).parseFromString(
-    // TODO do we want to support UTF-8?
-    String.fromCharCode.apply(null, new Uint16Array(message)),
-    'application/xml');
-  const headersElement = xml.getElementsByTagName('HttpHeaders')[0];
-  const headers = {};
-
-  if (headersElement) {
-    const headerNames = headersElement.getElementsByTagName('name');
-    const headerValues = headersElement.getElementsByTagName('value');
-
-    for (let i = 0; i < headerNames.length; i++) {
-      headers[headerNames[i].childNodes[0].nodeValue] =
-        headerValues[i].childNodes[0].nodeValue;
-    }
-  }
-
-  const challengeElement = xml.getElementsByTagName('Challenge')[0];
-  let challenge;
-
-  if (challengeElement) {
-    challenge = window.atob(challengeElement.childNodes[0].nodeValue);
-  }
-
-  return {
-    headers,
-    message: challenge
-  };
-};
 
 export const addKeyToSession = (options, session, event) => {
   let playreadyOptions = options.keySystems[PLAYREADY_KEY_SYSTEM];
@@ -55,15 +25,8 @@ export const addKeyToSession = (options, session, event) => {
   }
 
   const url = playreadyOptions.url || event.destinationURL;
-  const {headers, message} = getMessageContents(event.message.buffer);
 
-  videojs.xhr({
-    uri: url,
-    method: 'post',
-    headers,
-    body: message,
-    responseType: 'arraybuffer'
-  }, (err, response) => {
+  requestPlayreadyLicense(url, event.message.buffer, (err, response) => {
     if (err) {
       videojs.log.error('Unable to request key from url: ' + url);
       return;
