+++ src/plugin.js
@@ -57,6 +57,11 @@ const onPlayerReady = (player, options) => {
   });
   // IE11 Windows 8.1+
   player.tech_.el_.addEventListener('msneedkey', (event) => {
+    // EDGE still fires msneedkey, but should use encrypted instead
+    if (videojs.browser.IS_EDGE) {
+      return;
+    }
+
     handleMsNeedKeyEvent(event, videojs.mergeOptions(options, player.currentSource()));
   });
 };
